+++ src/main/java/de/samply/frontend/dto/DtoFactory.java
@@ -69,7 +69,9 @@ public class DtoFactory {
     public static User convert(@NotNull de.samply.db.model.ProjectBridgeheadUser projectBridgeheadUser) {
         return new User(
                 projectBridgeheadUser.getEmail(),
-                projectBridgeheadUser.getProjectBridgehead().getBridgehead()
+                projectBridgeheadUser.getProjectBridgehead().getBridgehead(),
+                projectBridgeheadUser.getProjectRole(),
+                projectBridgeheadUser.getProjectState()
         );
     }