+++ src/main/java/de/samply/frontend/dto/User.java
@@ -1,7 +1,12 @@
 package de.samply.frontend.dto;
 
+import de.samply.project.state.UserProjectState;
+import de.samply.user.roles.ProjectRole;
+
 public record User(
         String email,
-        String bridgehead
+        String bridgehead,
+        ProjectRole projectRole,
+        UserProjectState projectState
 ) {
 }